# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "floresu (floresu-external)",
    "version": "0.0.0"
  },
  "paths": {
...
    "/worklog": {
      "post": {
        "tags": [
          "worklog"
        ],
        "summary": "Create Entry",
        "operationId": "create_entry_worklog_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorklogWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorklogRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "worklog"
        ],
        "summary": "List Entries",
        "operationId": "list_entries_worklog_get",
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorklogSummary"
                  },
                  "title": "Response List Entries Worklog Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/worklog/{worklog_id}": {
      "get": {
        "tags": [
          "worklog"
        ],
        "summary": "Get Entry",
        "operationId": "get_entry_worklog__worklog_id__get",
        "parameters": [
          {
            "name": "worklog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worklog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorklogRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "worklog"
        ],
        "summary": "Update Entry",
        "operationId": "update_entry_worklog__worklog_id__put",
        "parameters": [
          {
            "name": "worklog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worklog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorklogWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorklogRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/worklog/{worklog_id}/archive": {
      "post": {
        "tags": [
          "worklog"
        ],
        "summary": "Archive Entry",
        "operationId": "archive_entry_worklog__worklog_id__archive_post",
        "parameters": [
          {
            "name": "worklog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worklog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorklogRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/worklog/{worklog_id}/restore": {
      "post": {
        "tags": [
          "worklog"
        ],
        "summary": "Restore Entry",
        "operationId": "restore_entry_worklog__worklog_id__restore_post",
        "parameters": [
          {
            "name": "worklog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worklog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorklogRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorklogRecord": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "entry_date": {
            "type": "string",
            "format": "date",
            "title": "Entry Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "bullet_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Bullet Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "entry_date",
          "description",
          "tags",
          "source_ids",
          "archived_at",
          "bullet_ids"
        ],
        "title": "WorklogRecord",
        "description": "A single entry, adding the bulletpoints that frame it (provenance)."
      },
      "WorklogSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "entry_date": {
            "type": "string",
            "format": "date",
            "title": "Entry Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "entry_date",
          "description",
          "tags",
          "source_ids",
          "archived_at"
        ],
        "title": "WorklogSummary",
        "description": "A worklog entry with its tags and attached sources (the timeline row)."
      },
      "WorklogWrite": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "entry_date": {
            "type": "string",
            "format": "date",
            "title": "Entry Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "entry_date"
        ],
        "title": "WorklogWrite",
        "description": "The create/update body: required title + date; optional description, tags, sources."
      }
    }
  }
}